Revert "Merge remote-tracking branch 'dinesh/PluginBasedCommandlet_Feature' into gcviewer"
This reverts commit 3c0868e2468825477518660a97359896b7dd9b7f, reversing
changes made to 4b6ae39dad2913be772c5e467721ae9ee37175a8.
